converting remaining files to Typescript
