+++ applications/jupyter-extension/nteract_on_jupyter/app/fonts.ts
@@ -1,5 +1,3 @@
-/* @flow strict */
-
 // TODO: There is the potential to have a Flash-of-Unstyled-Content since this is
 //       now async
 const WebFont = require("webfontloader");
